pyproject.toml
[tool.poetry]
name = "bundesliga-data-shootout"
version = "0.1.0"
description = ""
authors = ["Mykelo <michal.mieszko.wawrzyniak@gmail.com>"]
readme = "README.md"
packages = [{include = "bundesliga_data_shootout"}]

[tool.poetry.dependencies]
python = "^3.11"
opencv-python = "^4.7.0.72"
jupyter = "^1.0.0"
numpy = "^1.24.2"
pandas = "1.5.3"
matplotlib = "^3.7.1"
pytorch-lightning = "^2.0.1"
typer = "^0.7.0"
tqdm = "^4.65.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.3.0"
mypy = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
ignore_missing_imports = true
